Restructure definitions
