+++ services/graphql/src/definitions/index.js
@@ -1,22 +1,21 @@
 const { gql } = require('apollo-server');
+const company = require('./company');
+const leadership = require('./leadership');
 
 module.exports = gql`
 
-  scalar CompanyUpdateUpload
-  scalar CompanyUpdateDate
+  scalar Upload
+  scalar Date
 
   type Query {
+    ping: String!
     companyUpdateConfig: CompanyUpdateConfiguration!
-    companyUpdateSubmission(id: String!): CompanyUpdateSubmission!
     companyUpdateSubmissions(input: CompanyUpdateSubmissionsInput = {}): [CompanyUpdateSubmission!]!
     companyUpdateSubmissionCount: Int!
   }
 
   type Mutation {
-    companyUpdateSubmit(input: CompanyUpdateSubmissionInput!): CompanyUpdateSubmission!
-    companyUpdateComplete(id: String!): String!
-    companyUpdateDiscard(id: String!): String!
-    companyUpdateSingleUpload(file: CompanyUpdateUpload!): String!
+    ping: String!
   }
 
   enum ContactTypes {
@@ -36,66 +35,14 @@ module.exports = gql`
     logoUrl: String
     contactUrl: String
     leadershipSectionIds: [Int!]
-  }
-
-  type CompanyUpdateSubmission {
-    id: String!
-    submitted: CompanyUpdateDate!
-    reviewed: Boolean!
-    name: String!
-    hash: String!
-    email: String!
-    companyName: String
-    type: CompanyUpdateSubmissionType!
-    payload: String!
+    leadershipSectionMax: Int!
   }
 
   input CompanyUpdateSubmissionsInput {
     all: Boolean = false
   }
 
-  input CompanyUpdateSubmissionInput {
-    name: String!
-    email: String!
-    hash: String!
-    type: CompanyUpdateSubmissionType!
-    payload: CompanyUpdatePayloadInput!
-  }
-
-  input SocialLinkInput {
-    provider: String!
-    url: String!
-  }
-
-  input CompanyUpdatePayloadInput {
-    name: String
-    address1: String
-    address2: String
-    city: String
-    state: String
-    zip: String
-    country: String
-    phone: String
-    tollfree: String
-    fax: String
-    website: String
-    type: String
-    email: String
-    body: String
-    socialLinks: [SocialLinkInput!]
-
-    teaser: String
-    numberOfEmployees: String
-    trainingInformation: String
-    yearsInOperation: String
-    salesRegion: String
-    servicesProvided: String
-    salesChannels: String
-    productSummary: String
-    serviceInformation: String
-    warrantyInformation: String
-
-    logo: String
-  }
+  ${company}
+  ${leadership}
 
 `;
